did a bunch of stuff but forgot save
i did a bunch of commands. i got my circle to start at the very top and move downwards, when it gets to the bottom it resets at the top. i need to figure out how to make the circle spawn at a random spot on the top x axis